finalizando pratica 21
