+++ praticas/pomodoro/src/components/Footer/index.tsx
