+++ praticas/pomodoro/src/components/Logo/index.tsx
