+++ praticas/pomodoro/src/styles/globals.css
