--- praticas/pratica04/chronos-pomodoro/src/App.tsx
@@ -1,66 +1,0 @@
-import { Container } from './components/Container';
-import { Logo } from './components/Logo';
-import { Menu } from './components/Menu';
-import { CountDown } from './components/CountDown';
-import { DefaultInput } from './components/DefaultInput';
-import { Cycles } from './components/Cycles';
-import { DefaultButton } from './components/DefaultButton';
-import { PlayCircleIcon, StopCircleIcon } from 'lucide-react';
-import { Footer } from './components/Footer';
-
-import './styles/theme.css';
-import './styles/globals.css';
-
-export function App() {
-  return (
-    <>
-      <Container>
-        <Logo />
-      </Container>
-
-      <Container>
-        <Menu />
-      </Container>
-
-       <Container>
-        <CountDown />
-      </Container>
-
-      <Container>
-  <form className="form">
-  <div className="formRow">
-    <DefaultInput
-      id="meuInput"
-      type="text"
-      labelText="task"
-      placeholder="Digite algo"
-    />
-  </div>
-
-  <div className="formRow">
-    <p>Lorem ipsum dolor sit amet.</p>
-  </div>
-
-  <div className="formRow">
-    <Cycles />
-  </div>
-
-
-
-<div className="formRow">
-  <DefaultButton icon={<PlayCircleIcon />} />
-  <DefaultButton icon={<StopCircleIcon />} color="red" />
-</div>
-</form>
-</Container>
-
-<Container>
-  <Footer />
-</Container>
-
-</>
-  );
-}
-
-
-
